Epic3 frontend

+++ Server/WebAPI/Controllers/CommentController.cs
@@ -4,6 +4,7 @@ using BLL.Services;
 using DAL.Entities;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace WebAPI.Controllers
 {
@@ -52,15 +53,40 @@ namespace WebAPI.Controllers
         }
 
         [HttpDelete]
-        [Authorize(Roles = "Manager")]
+        [Authorize(Roles = "User, Manager, Admin")]
         [Route("{id:Guid}")]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-        public async Task<IActionResult> DeleteComment([FromBody] CommentModel model, [FromRoute] Guid id)
+        public async Task<IActionResult> DeleteComment([FromBody] CommentModel model,
+            [FromRoute] Guid id)
         {
-            model.GameId = id;
-            await _commentService.DeleteAsync(model);
-            return Ok("Deleted");
+            if (((!User.IsInRole("Manager") || (!User.IsInRole("Admin"))
+                    && model.UserId.ToString() ==
+                    User.Claims.First(u => u.Type == ClaimTypes.NameIdentifier).Value))
+                || (User.IsInRole("Admin") || User.IsInRole("Manager")))
+            {
+                model.GameId = id;
+                await _commentService.DeleteAsync(model);
+                return Ok(new { Response = "Deleted" });
+            }
+            return BadRequest(new { Response = "Do not have such permissions" });
+        }
+
+        [HttpDelete]
+        [Authorize(Roles = "User, Manager, Admin")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+        public async Task<IActionResult> DeleteComments([FromBody] CommentCollectionModel models)
+        {
+            if (User.IsInRole("Admin") || User.IsInRole("Manager"))
+            {
+                await _commentService.DeleteAsync(models.Comments);
+            }
+            else
+            {
+                await _commentService.DeleteAsync(models.Comments, User.Claims.First(u => u.Type == ClaimTypes.NameIdentifier).Value);
+            }
+            return Ok(new { Response = "Deleted" });
         }
     }
 }
